requirements.txt
pandas==2.0.3
PyYAML==6.0.1
scikit_learn==1.3.0
torch==2.2.1
torchtext==0.17.1
